requirements.txt
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
blessed==1.20.0
certifi==2025.4.26
Django==5.1.6
django-background-tasks==1.2.8
django-cors-headers==4.7.0
django-picklefield==3.3
django-q==1.3.9
geopandas==1.1.0
gunicorn==23.0.0
jinxed==1.3.0
numpy==2.2.6
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
pyogrio==0.11.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==3.5.3
shapely==2.1.1
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
tzdata==2025.1
wcwidth==0.2.13
whitenoise==6.9.0
